+++ src/sisl/_lib/_docscrape.py
